requirements.txt
Django==5.2.6
djangorestframework==3.16.1
Pillow==11.3.0
django-filter==25.1
requests==2.32.5
